various changes
- hotfix: should not re-run translations if nothing has changed
- switching language via the language selector now also switches language for translate nodes with `locale_is_selector` enabled
- removed lines in ChatNode and PineconeNode that were preventing templates loading in `/templates`
- the `templates` path now lays out the template and fits it to the view
